refactor: add tsPaths, move business logic from router
