+++ node_app/package.json
@@ -24,6 +24,8 @@
     "nodemon": "^2.0.13",
     "prettier": "^2.4.1",
     "ts-node": "^10.3.0",
+    "tsconfig-paths": "^3.11.0",
+    "tsconfig-paths-webpack-plugin": "^3.5.1",
     "typescript": "^4.4.4"
   },
   "dependencies": {
