+++ node_app/src/services/index.ts
@@ -1,0 +1,1 @@
+export * as userService from './userService';
