+++ node_app/src/services/userService.ts
@@ -1,4 +1,4 @@
-import { User } from './types';
+import { User } from '../types';
 
 export const users: User[] = [
   {
