+++ node_app/src/types/index.ts
@@ -1,3 +1,1 @@
-import { User } from './User';
-
-export { User };
+export { User } from './User';
